Moved most test cases and test macros back to `/intrin`.
