+++ src/arch/unknown/intrin/mod.rs
@@ -4,10 +4,16 @@ mod destride;
 mod downcast;
 mod endian;
 mod eq;
+mod hadd;
+mod hsub;
 mod merge;
 mod recip;
 mod round;
 mod rsqrt;
+mod saturating_add;
+mod saturating_hadd;
+mod saturating_sub;
+mod saturating_hsub;
 mod sum;
 mod sqrt;
 mod transmute;
@@ -20,10 +26,16 @@ pub mod prelude {
     pub use super::downcast::*;
     pub use super::endian::*;
     pub use super::eq::*;
+    pub use super::hadd::*;
+    pub use super::hsub::*;
     pub use super::merge::*;
     pub use super::recip::*;
     pub use super::round::*;
     pub use super::rsqrt::*;
+    pub use super::saturating_add::*;
+    pub use super::saturating_hadd::*;
+    pub use super::saturating_hsub::*;
+    pub use super::saturating_sub::*;
     pub use super::sum::*;
     pub use super::sqrt::*;
     pub use super::transmute::*;
