+++ src/arch/unknown/intrin/round.rs
@@ -44,192 +44,3 @@ rust_fallback_impl! {
         trunc => __undefined(_MM_FROUND_TRUNC), [0, 1, 2, 3];
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::prelude::*;
-
-    #[test]
-    fn round_f32s() {
-        let mut i = 0.0;
-        while i < 32.0 {
-            assert_eq!(f32s(i), f32s(i).round());
-            assert_eq!(f32s(i), f32s(i + 0.4).round());
-            assert_eq!(f32s(i + 1.0), f32s(i + 0.8).round());
-            i += 1.0;
-        }
-
-        let mut i = 0.0;
-        while i > -32.0 {
-            assert_eq!(f32s(i), f32s(i).round());
-            assert_eq!(f32s(i), f32s(i - 0.4).round());
-            assert_eq!(f32s(i - 1.0), f32s(i - 0.8).round());
-            i -= 1.0;
-        }
-
-        assert_eq!(f32s(9.8).round(), f32s(10.0));
-        assert_eq!(f32s(9.1).round(), f32s(9.0));
-        assert_eq!(f32s(9.0).round(), f32s(9.0));
-    }
-
-    #[test]
-    fn trunc_f32s() {
-        let mut i = 0.0;
-        while i < 32.0 {
-            assert_eq!(f32s(i), f32s(i).trunc());
-            assert_eq!(f32s(i), f32s(i + 0.4).trunc());
-            assert_eq!(f32s(i), f32s(i + 0.8).trunc());
-            i += 1.0;
-        }
-
-        let mut i = 0.0;
-        while i > -32.0 {
-            assert_eq!(f32s(i), f32s(i).trunc());
-            assert_eq!(f32s(i), f32s(i - 0.4).trunc());
-            assert_eq!(f32s(i), f32s(i - 0.8).trunc());
-            i -= 1.0;
-        }
-
-        assert_eq!(f32s(9.8).trunc(), f32s(9.0));
-        assert_eq!(f32s(9.1).trunc(), f32s(9.0));
-        assert_eq!(f32s(9.0).trunc(), f32s(9.0));
-    }
-
-    #[test]
-    fn ceil_f32s() {
-        let mut i = 0.0;
-        while i < 32.0 {
-            assert_eq!(f32s(i), f32s(i).ceil());
-            assert_eq!(f32s(i + 1.0), f32s(i + 0.4).ceil());
-            assert_eq!(f32s(i + 1.0), f32s(i + 0.8).ceil());
-            i += 1.0;
-        }
-
-        let mut i = 0.0;
-        while i > -32.0 {
-            assert_eq!(f32s(i), f32s(i).ceil());
-            assert_eq!(f32s(i), f32s(i - 0.4).ceil());
-            assert_eq!(f32s(i), f32s(i - 0.8).ceil());
-            i -= 1.0;
-        }
-
-        assert_eq!(f32s(9.8).ceil(), f32s(10.0));
-        assert_eq!(f32s(9.1).ceil(), f32s(10.0));
-        assert_eq!(f32s(9.0).ceil(), f32s(9.0));
-    }
-
-    #[test]
-    fn floor_f32s() {
-        let mut i = 0.0;
-        while i < 32.0 {
-            assert_eq!(f32s(i), f32s(i).floor());
-            assert_eq!(f32s(i), f32s(i + 0.4).floor());
-            assert_eq!(f32s(i), f32s(i + 0.8).floor());
-            i += 1.0;
-        }
-
-        let mut i = 0.0;
-        while i > -32.0 {
-            assert_eq!(f32s(i), f32s(i).floor());
-            assert_eq!(f32s(i - 1.0), f32s(i - 0.4).floor());
-            assert_eq!(f32s(i - 1.0), f32s(i - 0.8).floor());
-            i -= 1.0;
-        }
-
-        assert_eq!(f32s(9.8).floor(), f32s(9.0));
-        assert_eq!(f32s(9.1).floor(), f32s(9.0));
-        assert_eq!(f32s(9.0).floor(), f32s(9.0));
-    }
-
-    #[test]
-    fn round_f64s() {
-        let mut i = 0.0;
-        while i < 32.0 {
-            assert_eq!(f64s(i), f64s(i).round());
-            assert_eq!(f64s(i), f64s(i + 0.4).round());
-            assert_eq!(f64s(i + 1.0), f64s(i + 0.8).round());
-            i += 1.0;
-        }
-
-        let mut i = 0.0;
-        while i > -32.0 {
-            assert_eq!(f64s(i), f64s(i).round());
-            assert_eq!(f64s(i), f64s(i - 0.4).round());
-            assert_eq!(f64s(i - 1.0), f64s(i - 0.8).round());
-            i -= 1.0;
-        }
-
-        assert_eq!(f64s(9.8).round(), f64s(10.0));
-        assert_eq!(f64s(9.1).round(), f64s(9.0));
-        assert_eq!(f64s(9.0).round(), f64s(9.0));
-    }
-
-    #[test]
-    fn trunc_f64s() {
-        let mut i = 0.0;
-        while i < 32.0 {
-            assert_eq!(f64s(i), f64s(i).trunc());
-            assert_eq!(f64s(i), f64s(i + 0.4).trunc());
-            assert_eq!(f64s(i), f64s(i + 0.8).trunc());
-            i += 1.0;
-        }
-
-        let mut i = 0.0;
-        while i > -32.0 {
-            assert_eq!(f64s(i), f64s(i).trunc());
-            assert_eq!(f64s(i), f64s(i - 0.4).trunc());
-            assert_eq!(f64s(i), f64s(i - 0.8).trunc());
-            i -= 1.0;
-        }
-
-        assert_eq!(f64s(9.8).trunc(), f64s(9.0));
-        assert_eq!(f64s(9.1).trunc(), f64s(9.0));
-        assert_eq!(f64s(9.0).trunc(), f64s(9.0));
-    }
-
-    #[test]
-    fn ceil_f64s() {
-        let mut i = 0.0;
-        while i < 32.0 {
-            assert_eq!(f64s(i), f64s(i).ceil());
-            assert_eq!(f64s(i + 1.0), f64s(i + 0.4).ceil());
-            assert_eq!(f64s(i + 1.0), f64s(i + 0.8).ceil());
-            i += 1.0;
-        }
-
-        let mut i = 0.0;
-        while i > -32.0 {
-            assert_eq!(f64s(i), f64s(i).ceil());
-            assert_eq!(f64s(i), f64s(i - 0.4).ceil());
-            assert_eq!(f64s(i), f64s(i - 0.8).ceil());
-            i -= 1.0;
-        }
-
-        assert_eq!(f64s(9.8).ceil(), f64s(10.0));
-        assert_eq!(f64s(9.1).ceil(), f64s(10.0));
-        assert_eq!(f64s(9.0).ceil(), f64s(9.0));
-    }
-
-    #[test]
-    fn floor_f64s() {
-        let mut i = 0.0;
-        while i < 32.0 {
-            assert_eq!(f64s(i), f64s(i).floor());
-            assert_eq!(f64s(i), f64s(i + 0.4).floor());
-            assert_eq!(f64s(i), f64s(i + 0.8).floor());
-            i += 1.0;
-        }
-
-        let mut i = 0.0;
-        while i > -32.0 {
-            assert_eq!(f64s(i), f64s(i).floor());
-            assert_eq!(f64s(i - 1.0), f64s(i - 0.4).floor());
-            assert_eq!(f64s(i - 1.0), f64s(i - 0.8).floor());
-            i -= 1.0;
-        }
-
-        assert_eq!(f64s(9.8).floor(), f64s(9.0));
-        assert_eq!(f64s(9.1).floor(), f64s(9.0));
-        assert_eq!(f64s(9.0).floor(), f64s(9.0));
-    }
-}
