+++ src/intrin/sqrt.rs
@@ -22,3 +22,22 @@ pub trait Sqrt {
     fn sqrt(&self) -> Self;
 }
 
+
+#[cfg(test)]
+mod tests {
+    use crate::prelude::*;
+
+    #[test]
+    fn sqrt_f64s() {
+        assert_eq!(f64s(1.0).sqrt(), f64s(1.0));
+        assert!(f64s(9.0).sqrt().max(f64s(2.999)) == f64s(9.0).sqrt());
+        assert!(f64s(9.0).sqrt().min(f64s(3.001)) == f64s(9.0).sqrt());
+    }
+
+    #[test]
+    fn sqrt_f32s() {
+        assert_eq!(f32s(1.0).sqrt(), f32s(1.0));
+        assert!(f32s(9.0).sqrt().max(f32s(2.999)) == f32s(9.0).sqrt());
+        assert!(f32s(9.0).sqrt().min(f32s(3.001)) == f32s(9.0).sqrt());
+    }
+}
