+++ src/intrin/sum.rs
@@ -42,3 +42,39 @@ macro_rules! impl_packed_upcast_sum {
         )*
     }
 }
+
+
+macro_rules! test_packed_sum_int {
+    ($vec:tt, $el:tt, $name:ident) => {
+        #[test]
+        fn $name() {
+            // Try not to overflow
+            let mut i = $el::min_value() / 64 + 1;
+
+            while i < $el::max_value() / 64 - 1 {
+                let v = $vec::splat(i);
+                assert_eq!(v.sum(),
+                           v.scalar_reduce(0 as $el, |acc, v| acc + v));
+                assert_eq!(v.sum_upcast(),
+                           v.scalar_reduce(0 as i64, |acc, v| acc + (v as i64)));
+                i += $el::max_value() / 20;
+            }
+        }
+    };
+}
+
+macro_rules! test_packed_sum {
+    ($vec:tt, $el:tt, $name:ident) => {
+        #[test]
+        fn $name() {
+            for i in -100..100 {
+                let v = $vec::splat(i as $el);
+                assert_eq!(v.sum(),
+                           v.scalar_reduce(0 as $el, |acc, v| acc + v));
+                assert_eq!(v.sum_upcast(),
+                           v.scalar_reduce(0 as i64, |acc, v| acc + (v as i64)));
+            }
+        }
+    };
+}
+
