+++ src/lib.rs
@@ -215,6 +215,7 @@ pub mod into_iters;
 pub mod prelude;
 pub mod stride_zip;
 pub mod stride;
+
 pub use crate::prelude::*;
